fix(ai-hub): pin the Providers tab search + category dropdown on scroll
Mirror the Models tab: the search input and category select stick to the top of
the scroll container while the provider grid scrolls under them. Extracted the
Models tab's sentinel/stuck scroll-detection into a shared useStuckOnScroll hook
so both tabs share one implementation.
